Move reference synchronisation code to helper
